backend/requirements.txt
# FastAPI & Server
fastapi==0.115.8
uvicorn[standard]==0.34.0
gunicorn==23.0.0
python-multipart==0.0.20

# Pydantic v2
pydantic==2.10.6
pydantic-settings==2.7.1

# Azure SDKs
azure-identity==1.20.0
azure-keyvault-secrets==4.9.0
azure-keyvault-certificates==4.9.0
azure-cosmos==4.9.0
azure-storage-blob==12.24.1
openai==1.63.2
azure-ai-formrecognizer==3.3.3
azure-ai-contentsafety==1.0.0

# Microsoft Graph API
msal==1.32.0
httpx==0.28.1
aiohttp>=3.11.12

# Document Parsing
PyMuPDF==1.25.3
python-docx==1.1.2
Pillow>=10.0.0
opencv-python-headless>=4.8.0
reportlab>=4.2.5

# Web Crawling
playwright==1.50.0
beautifulsoup4==4.13.3
requests==2.32.3
lxml>=4.9.0
crawl4ai>=0.4.0
googlesearch-python>=1.3.0

# PII Masking - Presidio
# spaCy 3.8.x does not support Python 3.14 yet; use latest compatible build
spacy>=3.8.2
presidio-analyzer>=2.2.357
presidio-anonymizer>=2.2.357

# Encryption
cryptography==44.0.1

# Logging & Utils
python-json-logger==3.2.1
python-dotenv==1.0.1
tenacity==9.0.0
 
# MongoDB Support (Cosmos DB Mongo API)
motor==3.6.1
dnspython==2.7.0
pyopenssl>=23.2.0
 
# Demo DB
tinydb==4.8.2

# JWT validation
PyJWT==2.10.1

# Pipeline V2 — new dependencies
rapidfuzz>=3.6.0
trafilatura>=1.9.0

# Testing
pytest==8.3.5
pytest-asyncio==0.25.3
